raft: rename configuration.h to group_configuration.h
Have the file name reflecting the class name

Fixes #505
