+++ src/v/cluster/partition.h
@@ -15,8 +15,8 @@
 #include "cluster/types.h"
 #include "model/metadata.h"
 #include "model/record_batch_reader.h"
-#include "raft/configuration.h"
 #include "raft/consensus.h"
+#include "raft/group_configuration.h"
 #include "raft/id_allocator_stm.h"
 #include "raft/log_eviction_stm.h"
 #include "raft/types.h"
