+++ src/v/raft/consensus.cc
@@ -13,10 +13,10 @@
 #include "likely.h"
 #include "model/metadata.h"
 #include "prometheus/prometheus_sanitize.h"
-#include "raft/configuration.h"
 #include "raft/consensus_client_protocol.h"
 #include "raft/consensus_utils.h"
 #include "raft/errc.h"
+#include "raft/group_configuration.h"
 #include "raft/logger.h"
 #include "raft/prevote_stm.h"
 #include "raft/recovery_stm.h"
