+++ src/v/raft/consensus.h
@@ -15,11 +15,11 @@
 #include "model/fundamental.h"
 #include "model/metadata.h"
 #include "raft/append_entries_buffer.h"
-#include "raft/configuration.h"
 #include "raft/configuration_manager.h"
 #include "raft/consensus_client_protocol.h"
 #include "raft/event_manager.h"
 #include "raft/follower_stats.h"
+#include "raft/group_configuration.h"
 #include "raft/logger.h"
 #include "raft/mutex_buffer.h"
 #include "raft/prevote_stm.h"
