+++ src/v/raft/group_configuration.h
