+++ src/v/raft/heartbeat_manager.cc
@@ -13,9 +13,9 @@
 #include "model/metadata.h"
 #include "model/timeout_clock.h"
 #include "outcome_future_utils.h"
-#include "raft/configuration.h"
 #include "raft/consensus_client_protocol.h"
 #include "raft/errc.h"
+#include "raft/group_configuration.h"
 #include "raft/raftgen_service.h"
 #include "raft/types.h"
 #include "rpc/reconnect_transport.h"
