+++ src/v/raft/replicate_entries_stm.cc
@@ -13,10 +13,10 @@
 #include "model/fundamental.h"
 #include "model/metadata.h"
 #include "outcome_future_utils.h"
-#include "raft/configuration.h"
 #include "raft/consensus.h"
 #include "raft/consensus_utils.h"
 #include "raft/errc.h"
+#include "raft/group_configuration.h"
 #include "raft/logger.h"
 #include "raft/raftgen_service.h"
 #include "raft/types.h"
