+++ src/v/raft/tests/foreign_entry_test.cc
@@ -13,9 +13,9 @@
 #include "model/record.h"
 #include "model/record_batch_reader.h"
 #include "model/timeout_clock.h"
-#include "raft/configuration.h"
 #include "raft/configuration_bootstrap_state.h"
 #include "raft/consensus_utils.h"
+#include "raft/group_configuration.h"
 #include "raft/types.h"
 #include "random/generators.h"
 #include "resource_mgmt/io_priority.h"
