+++ src/v/raft/tests/type_serialization_tests.cc
@@ -12,8 +12,8 @@
 #include "model/record.h"
 #include "model/record_batch_reader.h"
 #include "model/timeout_clock.h"
-#include "raft/configuration.h"
 #include "raft/consensus_utils.h"
+#include "raft/group_configuration.h"
 #include "raft/types.h"
 #include "random/generators.h"
 #include "storage/record_batch_builder.h"
