+++ src/v/raft/types.cc
@@ -11,8 +11,8 @@
 
 #include "model/fundamental.h"
 #include "model/metadata.h"
-#include "raft/configuration.h"
 #include "raft/consensus_utils.h"
+#include "raft/group_configuration.h"
 #include "reflection/adl.h"
 #include "utils/to_string.h"
 #include "vassert.h"
